feat(net)!: remove service proxy and service provider
