+++ cellularjs.com/docs/3. foundation/1. net/3. service.md
@@ -75,111 +75,3 @@ export class Task implements ServiceHandler {
 }
 ```
 :::
-
-## 4. Service helpers
-:::danger __DEPRECATED__
-[Service proxy](/docs/foundation/net/service#41-service-proxy) and [service provider](/docs/foundation/net/service#42-service-providers) will be removed in the next version(0.12.0). You can use [proxy](/docs/foundation/dependency-injection/basic-usage#3-proxy) instead.
-:::
-
-### 4.1. Service proxy
-Service proxy is a type of `ServiceHandler`, it help you modify/add logic to a specific service. You can use `addServiceProxies` to register proxies for a service.
-
-:::info What is `NextHandler`?
-You can consider it as `next` function in `express`.
-Instead of resolve all handlers/proxies and it's dependencies at once, `NextHandler` allow you to resolve and run next (proxy/service) handler lazily.
-:::
-
-Example: create simple `AuthGuard` decorator.
-```ts {24}
-import { Injectable } from '@cellularjs/di';
-import { Service, ServiceHandler, NextHandler, addServiceProxies, IRS, IRQ } from '@cellularjs/net';
-
-@Injectable()
-class AuthGuardProxy implements ServiceHandler {
-  constructor(
-    private irq: IRQ,
-    private nextHandler: NextHandler,
-  ) { }
-
-  handle() {
-    const { irq, nextHandler } = this;
-
-    if (!irq.header.accessToken) {
-      throw new IRS({ status: 401 }, { err: 'UNAUTHORIZED' });
-    }
-
-    // It work like a pipeline of service handlers
-    return nextHandler.handle();
-  }
-}
-
-const AuthGuard = () => service => {
-  addServiceProxies(service, [AuthGuardProxy]);
-
-  return service;
-}
-
-@AuthGuard()
-@Service({ scope: 'publish' })
-export class CreatePost {
-  handle() {
-    // ...
-  }
-}
-```
-
-### 4.2. Service providers
-In addition to [providers](/docs/foundation/net/cell#21-providers) configured in `Cell` decorator, you can add more providers to a specific service with `addServiceProviders`.
-
-:::tip You should known
-Service providers is very different from [normal providers](/docs/foundation/net/cell#21-providers). Service providers will be resolved on every internal request. That mean `permanent` provider is only "permanent" in a single request!
-:::
-
-Example: create a dummy `AuthGuard` decorator that support role based access control.
-
-```ts {31-33}
-import { Injectable, Inject } from '@cellularjs/di';
-import { Service, ServiceHandler, NextHandler, addServiceProxies, addServiceProviders, IRS, IRQ } from '@cellularjs/net';
-
-@Injectable()
-class AuthGuardProxy implements ServiceHandler {
-  constructor(
-    private irq: IRQ,
-    private nextHandler: NextHandler,
-    @Inject('allowedRoles')
-    private allowedRoles: string[],
-  ) { }
-
-  handle() {
-    const { irq, nextHandler, allowedRoles } = this;
-
-    if (!irq.header.accessToken) {
-      throw new IRS({ status: 401 }, { err: 'UNAUTHORIZED' });
-    }
-
-    const user = { role: 'guest' };
-    if (allowedRoles.includes(user.role)) {
-      throw new IRS({ status: 403 }, { err: 'ACCESS_FORBIDDEN' });
-    }
-
-    return nextHandler.handle();
-  }
-}
-
-const AuthGuard = (allowedRoles) => service => {
-  addServiceProxies(service, [AuthGuardProxy]);
-  addServiceProviders(service, [
-    { token: 'allowedRoles', useValue: allowedRoles },
-  ]);
-
-  return service;
-}
-
-@AuthGuard(['admin'])
-@Service({ scope: 'publish' })
-export class CreatePost {
-  handle() {
-    // ...
-  }
-}
-```
