+++ cellularjs.com/docs/3. foundation/1. net/6. transporter.md
@@ -219,7 +219,7 @@ Whenever you make an internal request, CellularJS can emit these events: `start`
 :::caution
 `transportListener` is a *GLOBAL* object. That mean listener will be able to listen and change data of every requests from different cells. So it is only suitable for "global" tasks such as logging request,... and code should be used at common place(Eg: `$share` folder).
 
-**_If you want to modify/add behaviour to specific service, have a look at [service proxy](/docs/foundation/net/service#41-service-proxy)._**
+**_If you want to modify/add behaviour to specific service, have a look at [proxy](/docs/foundation/dependency-injection/basic-usage#3-proxy)._**
 :::
 
 :::note
