+++ cellularjs.com/docs/3. foundation/1. net/7. api.md
@@ -29,22 +29,7 @@ createNetWork(networkConfig: NetworkConfig): Promise<void[]>;
 getResolvedCell(cellName: string): ResolvedCell | undefined;
 ```
 
-## 3. Service-related APIs
-```ts
-/**
- * @see https://cellularjs.com/docs/foundation/net/service/#41-service-proxy
- */
-addServiceProxies(service, proxies: ClassType<ServiceHandler>[]): void;
-```
-
-```ts
-/**
- * @see https://cellularjs.com/docs/foundation/net/service#42-service-providers
- */
-addServiceProviders(service, providers: GenericProvider[]): void;
-```
-
-## 4. Transporter-related APIs
+## 3. Transporter-related APIs
 ```ts
 /**
  * Send Internal Request and get Internal Response.
@@ -60,12 +45,12 @@ send(irq: IRQ, requestOpts?: RequestOptions): Promise<IRS>;
 const transportListener: TransportListener;
 ```
 
-## 5. Types
+## 4. Types
 
-### 5.1. NetworkConfig
+### 4.1. NetworkConfig
 **`NetworkConfig = CellConfig[];`**
 
-### 5.2. CellConfig
+### 4.2. CellConfig
 | Name       | Type                                           | Required | Default | Description                                                                                                                                               |
 |:-----------|:-----------------------------------------------|:---------|:--------|-----------------------------------------------------------------------------------------------------------------------------------------------------------|
 | name       | string                                         | True     |         | Cell name or cell type, it must be unique.                                                                                                                |
@@ -73,19 +58,19 @@ const transportListener: TransportListener;
 | driver     | { new() }; { [driverName: string]: { new() } } | True     |         | Driver is a thing that handle request, [read more...](/docs/foundation/net/virtual-network#13-driver).                                                    |
 | customData | { [key: string]: any }                         | False    |         | In additional to reserved property defined by `CellConfig`, you can add more custom data with this property. You can get this data via `getResolvedCell`. |
 
-### 5.3. CellMeta
+### 4.3. CellMeta
 | Name      | Type                                      | Required | Default | Description                                                                                                                                                    |
 |:----------|:------------------------------------------|:---------|:--------|----------------------------------------------------------------------------------------------------------------------------------------------------------------|
 | providers | (GenericProvider, string)[]               | False    |         | Declare providers for resolving dependency later. For more information, have a look at [DI API](/docs/foundation/dependency-injection/api#31-genericprovider). |
 | imports   | ImportableCnf[]                           | False    |         | Import module into current cell.                                                                                                                               |
 | listen    | string, { [key: string]: ServiceHandler } | True     |         | Register services into cell, [read more...](/docs/foundation/net/cell#23-listen).                                                                              |
 
-### 5.4. ServiceMeta
+### 4.4. ServiceMeta
 | Name  | Type                          | Required | Default | Description                                                                                                                                                                                |
 |:------|:------------------------------|:---------|:--------|:-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|
 | scope | 'publish', 'space', 'private' | False    | 'space' | Service handler access scope:<br/>- 'publish': accessible from anywhere.<br/>- 'space': limit access to the cell with the same space only.<br/>- 'private': limit access to the cell only. |
 
-### 5.5. RequestOptions
+### 4.5. RequestOptions
 | Name               | Type    | Required | Default           | Description                                                                                                                                                                             |
 |--------------------|---------|----------|-------------------|-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|
 | driver             | string  | False    | 'local'           | Driver to handle this request, [read more...](/docs/foundation/net/virtual-network#13-driver).                                                                                          |
