+++ packages/net/src/internal.ts
@@ -1,9 +1,8 @@
 export * from './cell-context';
 export * from './error';
-export * from './service-helper';
 export * from './type';
 export * from './key';
 export * from './decorator';
 export * from './network';
-export * from './transportor';
+export * from './transporter';
 export * from './message';
