+++ packages/net/src/network/clean-network.func.ts
@@ -1,8 +1,5 @@
-import { cleanServiceProviders, cleanServiceProxies } from '../service-helper';
 import { cleanResolvedCells } from './resolved-cell.data';
 
 export async function cleanNetwork() {
   cleanResolvedCells();
-  cleanServiceProviders();
-  cleanServiceProxies();
 }
