--- packages/net/src/service-helper/index.ts
@@ -1,10 +1,0 @@
-export {
-  addServiceProviders,
-  cleanServiceProviders,
-  getServiceProviders,
-} from './service-provider';
-export {
-  addServiceProxies,
-  cleanServiceProxies,
-  getServiceProxies,
-} from './service-proxy';
