+++ packages/net/src/transporter/index.ts
@@ -1,3 +1,2 @@
 export { send } from './send.func';
 export { transportListener, TransportListener } from './transport-emitter';
-export { NextHandler } from './next-handler';
